Fix some warnings on Windows.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1478553 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/core/kernels/ActiveMQConsumerKernel.cpp
@@ -298,7 +298,7 @@ namespace kernels {
 
         void waitForRedeliveries() {
             if (failoverRedeliveryWaitPeriod > 0 && previouslyDeliveredMessages != NULL) {
-                long expiry = System::currentTimeMillis() + failoverRedeliveryWaitPeriod;
+                long long expiry = System::currentTimeMillis() + failoverRedeliveryWaitPeriod;
                 int numberNotReplayed;
                 do {
                     numberNotReplayed = 0;
